Update load.rb file to load games and label data
