+++ json/authors.json
@@ -1,0 +1,34 @@
+[
+  {
+    "id": 715,
+    "author": [
+      {
+        "first name": "John",
+        "last name": "Legend"
+      }
+    ],
+    "items": [
+      {
+        "publish_date": null,
+        "multiplayer": false,
+        "last_played_at": "2020-04-04"
+      }
+    ]
+  },
+  {
+    "id": 21,
+    "author": [
+      {
+        "first name": "Steve",
+        "last name": "Opoku"
+      }
+    ],
+    "items": [
+      {
+        "publish_date": null,
+        "multiplayer": false,
+        "last_played_at": "2009-03-05"
+      }
+    ]
+  }
+]
