+++ json/labels.json
@@ -2,5 +2,9 @@
   {
     "title": "Hope",
     "color": "red"
+  },
+  {
+    "title": "Harry potter",
+    "color": "green"
   }
 ]
